requirements.txt
streamlit==1.27.0
psycopg2-binary==2.9.7
